Quick refacto
