+++ src/lib.rs
@@ -1,3 +1,4 @@
-pub mod oracle;
-pub mod utils;
 pub mod conversions;
+pub mod oracle;
+pub mod types;
+pub mod utils;
